backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
dj-database-url==3.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==23.0.0
whitenoise==6.9.0
openai==2.21.0
celery==5.6.2
redis==5.3.0
requests==2.32.5
pydantic==2.12.5
httpx==0.28.1
